publisher/requirements.txt
tweepy==3.8.0
kafka-python==2.0.1
pytest==5.4.1